C: Move `hb_arena.c` to `src/util/`
