+++ javascript/packages/node/binding.gyp
@@ -25,7 +25,6 @@
         "./extension/libherb/lexer_peek_helpers.c",
         "./extension/libherb/lexer.c",
         "./extension/libherb/location.c",
-        "./extension/libherb/hb_arena.c",
         "./extension/libherb/parser_helpers.c",
         "./extension/libherb/parser.c",
         "./extension/libherb/pretty_print.c",
@@ -35,6 +34,7 @@
         "./extension/libherb/token.c",
         "./extension/libherb/utf8.c",
         "./extension/libherb/util.c",
+        "./extension/libherb/util/hb_arena.c",
         "./extension/libherb/util/hb_array.c",
         "./extension/libherb/util/hb_buffer.c",
         "./extension/libherb/util/hb_string.c",
